+++ GJL Jam 2022/Assets/Scripts/Enemies/Minions/MinionStats.cs
@@ -5,11 +5,11 @@ using UnityEngine;
 public struct MinionStats
 {
     public int _level;
-    public int _damage;
+    public float _damage;
     public float _speed;
     public float _size;
 
-    public MinionStats(int level, int damage, float speed, float size)
+    public MinionStats(int level, float damage, float speed, float size)
     {
         _level = level;
         _damage = damage;
